// src/Garage-TraeUI/backend/src/types.ts
export interface BucketPermissions {
  read: boolean;
  write: boolean;
  owner: boolean;
}
// ...
export interface AccessKeyPermissions {
  create_bucket: boolean;
}
// ...
export interface CreateAccessKeyRequest {
  name?: string;
}

export interface UpdateAccessKeyRequest {
  name?: string;
  permissions?: AccessKeyPermissions;
  bucket_aliases?: string[];
  bucket_permissions?: Record<string, BucketPermissions>;
}